requirements.txt
cloudscraper==1.2.68
ffmpeg-python==0.2.0
beautifulsoup4==4.12.2
tqdm==4.66.3
python-dotenv==1.0.0
requests==2.32.0
pyinstaller==6.12.0
chardet==4.0.0
inquirer==3.4.0
playwright==1.51.0